+++ services/cosmos-db/mgmt/2021-10-15/documentdb/CHANGELOG.md
@@ -1,2 +1,7 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Constants
+
+1. ServerVersion.ServerVersionFourFullStopTwo