+++ services/preview/cosmos-db/mgmt/2021-11-15-preview/documentdb/CHANGELOG.md
@@ -1,2 +1,9 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. KeyWrapMetadata.Algorithm